--- backEnd/api/controllers/GetAllViewFilters.js
@@ -1,10 +1,0 @@
-const ViewFilters = require('../models/viewFilterModel');
-
-module.exports = {
-  GetAllViewFilters:GetAllViewFilters
-}
-
-async function GetAllViewFilters(req,res){
-  let allViewFilters = await ViewFilters.find();
-  res.json(allViewFilters);
-}
